Moving jetpack_is_mobile into a package [not verified]
