+++ composer.json
@@ -22,6 +22,7 @@
 		"automattic/jetpack-error": "@dev",
 		"automattic/jetpack-jitm": "@dev",
 		"automattic/jetpack-logo": "@dev",
+		"automattic/jetpack-mobile": "@dev",
 		"automattic/jetpack-options": "@dev",
 		"automattic/jetpack-partner": "@dev",
 		"automattic/jetpack-redirect": "@dev",
