+++ packages/mobile/.gitattributes
@@ -1,0 +1,4 @@
+# Files not needed to be distributed in the package.
+.gitattributes   export-ignore
+phpunit.xml.dist export-ignore
+tests/           export-ignore
