+++ packages/mobile/tests/php/test-mobile.php
@@ -1,0 +1,8 @@
+<?php
+
+use Automattic\Jetpack\Constants;
+use phpmock\Mock;
+use phpmock\spy\Spy;
+use PHPUnit\Framework\TestCase;
+
+class Test_Mobile extends TestCase {}
